Revert "[Enhancement] support cast struct/map/array to json (#19476)"
This reverts commit 466487a30cb9d3f9c51ea1ecdba6ae38ea2ed7fb.
