+++ be/src/exprs/cast_expr.h
@@ -14,7 +14,6 @@
 
 #pragma once
 
-#include <type_traits>
 #include <utility>
 
 #include "column/array_column.h"
@@ -24,7 +23,6 @@
 #include "column/vectorized_fwd.h"
 #include "exprs/column_ref.h"
 #include "exprs/expr.h"
-#include "runtime/large_int_value.h"
 #include "runtime/types.h"
 
 namespace starrocks {
@@ -150,31 +148,5 @@ public:
 private:
     std::vector<std::unique_ptr<Expr>> _field_casts;
 };
-
-/**
- * Cast other type to string without float, double, string
- */
-struct CastToString {
-    template <typename Type, typename ResultType>
-    static std::string apply(const Type& v) {
-        if constexpr (IsTemporal<Type>() || IsDecimal<Type>) {
-            // DateValue, TimestampValue, DecimalV2
-            return v.to_string();
-        } else if constexpr (IsInt128<Type>) {
-            // int128_t
-            return LargeIntValue::to_string(v);
-        } else {
-            // int8_t ~ int64_t, boolean
-            return SimpleItoa(v);
-        }
-    }
-
-    template <typename Type>
-    static bool constexpr extend_type() {
-        return (IsTemporal<Type>() || IsDecimal<Type> || IsInt128<Type>);
-    }
-};
-
-StatusOr<ColumnPtr> cast_nested_to_json(const ColumnPtr& column);
 
 } // namespace starrocks
