+++ fe/fe-core/src/main/java/com/starrocks/catalog/PrimitiveType.java
@@ -109,19 +109,18 @@ public enum PrimitiveType {
     public static final ImmutableList<PrimitiveType> STRING_TYPE_LIST =
             ImmutableList.of(CHAR, VARCHAR);
 
-    private static final ImmutableList<PrimitiveType> TIME_TYPE_LIST =
-            ImmutableList.of(TIME, DATE, DATETIME);
-
     public static final ImmutableList<PrimitiveType> JSON_COMPATIBLE_TYPE =
             new ImmutableList.Builder<PrimitiveType>()
                     .add(BOOLEAN)
-                    .add(DATE)
-                    .add(DATETIME)
                     .addAll(NUMBER_TYPE_LIST)
                     .addAll(STRING_TYPE_LIST)
                     .build();
+    // TODO(mofei) support them
     public static final ImmutableList<PrimitiveType> JSON_UNCOMPATIBLE_TYPE =
-            ImmutableList.of(TIME, HLL, BITMAP, PERCENTILE, FUNCTION, VARBINARY);
+            ImmutableList.of(DATE, DATETIME, TIME, HLL, BITMAP, PERCENTILE, FUNCTION, VARBINARY);
+
+    private static final ImmutableList<PrimitiveType> TIME_TYPE_LIST =
+            ImmutableList.of(TIME, DATE, DATETIME);
 
     private static final ImmutableList<PrimitiveType> BINARY_TYPE_LIST = ImmutableList.of(VARBINARY);
