+++ fe/fe-core/src/main/java/com/starrocks/catalog/Type.java
@@ -1084,8 +1084,6 @@ public abstract class Type implements Cloneable {
         } else if (from.isJsonType() && to.isArrayScalar()) {
             // now we only support cast json to one dimensional array
             return true;
-        } else if ((from.isStructType() || from.isMapType()) && to.isJsonType()) {
-            return true;
         } else {
             return false;
         }
