+++ fe/fe-core/src/test/java/com/starrocks/sql/plan/JsonTypeTest.java
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+
 package com.starrocks.sql.plan;
 
 import com.starrocks.common.ExceptionChecker;
@@ -42,8 +43,6 @@ public class JsonTypeTest extends PlanTestBase {
                 " v_FLOAT FLOAT," +
                 " v_VARCHAR VARCHAR," +
                 " v_CHAR CHAR," +
-                " v_DATE DATE," +
-                " v_DATETIME DATETIME, " +
                 " v_decimal decimal128(32,1)" +
                 ") DUPLICATE KEY (v_id) " +
                 "DISTRIBUTED BY HASH (v_id) " +
@@ -168,16 +167,26 @@ public class JsonTypeTest extends PlanTestBase {
             assertPlanContains("select cast(v_json as DECIMAL128(32,1)) from tjson_test",
                     "CAST(2: v_json AS DECIMAL128(32,1))");
             assertPlanContains("select cast(v_decimal AS JSON) from tjson_test",
-                    "CAST(15: v_decimal AS JSON)");
+                    "CAST(13: v_decimal AS JSON)");
         }
 
-        {
-            // Temporal types
-            List<String> temporalTypeList = Arrays.asList("date", "datetime");
-            for (String temporalType : temporalTypeList) {
-                String columnCastSql = String.format("select cast(v_%s as JSON) from tjson_test", temporalType);
-                getFragmentPlan(columnCastSql);
-            }
+        List<String> notAllowedCastTypes = Arrays.asList("date", "datetime");
+        for (String notAllowType : notAllowedCastTypes) {
+            String columnCastSql = String.format("select cast(v_json as %s) from tjson_test", notAllowType);
+            ExceptionChecker.expectThrowsWithMsg(
+                    SemanticException.class,
+                    String.format(
+                            "Invalid type cast from json to %s in sql `test.tjson_test.v_json`",
+                            notAllowType),
+                    () -> getFragmentPlan(columnCastSql)
+            );
+
+            String functionCastSql = String.format("select cast(parse_json('') as %s)", notAllowType);
+            ExceptionChecker.expectThrowsWithMsg(
+                    SemanticException.class,
+                    String.format("Invalid type cast from json to %s in sql `parse_json('')`", notAllowType),
+                    () -> getFragmentPlan(functionCastSql)
+            );
         }
 
     }
@@ -225,10 +234,5 @@ public class JsonTypeTest extends PlanTestBase {
         assertExceptionMessage("select cast(json_array(1,2,3) as array<array<int>>)",
                 "Getting analyzing error from line 1, column 7 to line 1, column 50. Detail message: " +
                         "Invalid type cast from json to ARRAY<ARRAY<int(11)>> in sql `json_array(1, 2, 3)`.");
-    }
-
-    @Test
-    public void testCastStructToJson() throws Exception {
-        assertPlanContains("select cast(row(1, 2) as json)", "CAST(row(1, 2) AS JSON)");
     }
 }
